// src/api/generated/models/SeleccionDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SeleccionDTO = {
    id: number;
    nombre: string;
    pais: string;
    grupo: string;
    bandera?: string;
    partidosJugados: number;
    partidosGanados: number;
    partidosEmpatados: number;
    partidosPerdidos: number;
    golesAFavor: number;
    golesEnContra: number;
    puntos: number;
    diferenciaGoles: number;
};